Send only used log records to MP Logging Listener
Currently, the architecture sends all log records to the Multiprocessing
Logging Listener through a multiprocessing queue. On production systems,
this will send many records that ultimately will not be used, including
debug and trace level log records from Salt and any dependencies of Salt
and its plugins. Such log records are numerous and some may be very large
due to containing packet information. This creates an unneeded burden on a
production system.

This change will determine the minimum needed logging level based on
the configuration and use that for the Multiprocessing Logging clients.

Signed-off-by: Sergey Kizunov <sergey.kizunov@ni.com>